Enforce invariant: (LAIMAX==0) == (LC==0)
If maximum LAI doesn't exist but there's a land cover type there, then that's a problem.  Should spit out an error message.  I've got this type of vegetation here but it never has any leaves.  That means classification inconsistency between MODIS LAI and cover products.  We should really spit out an error map (by PFT)... these are all the grid cells where there is no LAI but there is this vegetation type.  Thing to do would be to set it to 0.  It's this cover that never has an LAI.  But accumulate a map for all points for each PFT where cover and LAI are inconsistent.

laic is max. LAI.  This cover only does MAX LAI and cover.  It takes cover and max LAI, and assigns LAI to the cover.  Before this program, laimax is generated by cover type.  (It was done in A01... that program should have produced the error maps and checked for inconsistencies).